# swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TodoListApp API",
    "version": "1.0.0",
    "description": "TodoListApp MVP REST API specification. JWT is issued by the backend and stored only in the frontend Zustand in-memory store. JWT must not be stored in localStorage, sessionStorage, or cookies."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development API server"
    }
  ],
  "tags": [
    {
      "name": "Auth",
      "description": "회원가입, 로그인, 로그아웃"
    },
    {
      "name": "Users",
      "description": "내 정보 조회, 수정, 회원 탈퇴"
    },
    {
      "name": "Categories",
      "description": "기본 카테고리 및 사용자 추가 카테고리"
    },
    {
      "name": "Todos",
      "description": "?? CRUD, ?? ??, ???"
    },
    {
      "name": "System",
      "description": "서버 상태 확인"
    }
  ],
  "paths": {
    "/auth/signup": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "회원가입",
        "description": "이메일, 비밀번호, 이름 또는 닉네임으로 사용자 계정을 생성한다.",
        "operationId": "signup",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              },
              "examples": {
                "default": {
                  "value": {
                    "email": "user@example.com",
                    "password": "password123!",
                    "displayName": "김민준"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "회원가입 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "로그인",
        "description": "이메일과 비밀번호를 검증하고 JWT를 발급한다.",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "examples": {
                "default": {
                  "value": {
                    "email": "user@example.com",
                    "password": "password123!"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "로그인 성공. 클라이언트는 token을 Zustand 메모리 상태에만 저장한다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "로그아웃",
        "description": "MVP에서는 서버 측 JWT 블랙리스트를 사용하지 않는다. 이 API는 서버 상태를 변경하지 않는 no-op이며, 클라이언트가 Zustand 메모리 상태의 JWT를 삭제한다. JWT가 없거나 만료되어도 프론트엔드는 메모리 인증 상태를 삭제해야 한다.",
        "operationId": "logout",
        "responses": {
          "204": {
            "description": "로그아웃 처리 완료"
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "헬스 체크",
        "description": "개발 및 운영 확인을 위한 서버 상태 확인 API이다.",
        "operationId": "healthCheck",
        "responses": {
          "200": {
            "description": "서버 정상",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "내 정보 조회",
        "description": "인증된 사용자의 이메일과 이름 또는 닉네임을 조회한다.",
        "operationId": "getMe",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "내 정보 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "내 정보 수정",
        "description": "MVP에서는 이름 또는 닉네임, 다크모드 사용 여부, 화면 표시 언어를 수정한다. 이메일과 비밀번호 변경은 제공하지 않는다.",
        "operationId": "updateMe",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "내 정보 수정 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "회원 탈퇴",
        "description": "사용자 계정, 본인 소유 할일, 사용자 추가 카테고리를 즉시 삭제한다.",
        "operationId": "deleteMe",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "회원 탈퇴 성공"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/categories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "카테고리 목록 조회",
        "description": "기본 카테고리와 현재 사용자가 생성한 사용자 추가 카테고리를 조회한다.",
        "operationId": "listCategories",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "카테고리 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Category"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "summary": "사용자 카테고리 생성",
        "description": "현재 사용자에게만 귀속되는 사용자 추가 카테고리를 생성한다.",
        "operationId": "createCategory",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "카테고리 생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/todos": {
      "get": {
        "tags": [
          "Todos"
        ],
        "summary": "할일 목록 조회",
        "description": "?? ??? ?? ??? ????. ????, ????? ??, ?? ??? ??? ? ??. MVP??? ??????? ???? ???.",
        "operationId": "listTodos",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "카테고리 ID"
          },
          {
            "name": "dueDateFrom",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "종료예정일 시작일"
          },
          {
            "name": "dueDateTo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "종료예정일 종료일"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "registered",
                "in_progress",
                "completed",
                "canceled"
              ]
            },
            "description": "?? ??. ???? ?? ??"
          }
        ],
        "responses": {
          "200": {
            "description": "할일 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Todos"
        ],
        "summary": "할일 등록",
        "description": "할일명, 설명, 종료예정일, 카테고리를 입력해 현재 사용자 소유 할일을 생성한다.",
        "operationId": "createTodo",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTodoRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "할일 등록 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/todos/{todoId}": {
      "get": {
        "tags": [
          "Todos"
        ],
        "summary": "할일 상세 조회",
        "description": "현재 사용자 소유의 특정 할일을 조회한다.",
        "operationId": "getTodo",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TodoId"
          }
        ],
        "responses": {
          "200": {
            "description": "할일 상세 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "tags": [
          "Todos"
        ],
        "summary": "할일 수정",
        "description": "?? ??? ??? ???, ??, ?????, ????, ??? ????.",
        "operationId": "updateTodo",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TodoId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTodoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "할일 수정 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "Todos"
        ],
        "summary": "할일 삭제",
        "description": "현재 사용자 소유의 할일을 삭제한다.",
        "operationId": "deleteTodo",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TodoId"
          }
        ],
        "responses": {
          "204": {
            "description": "할일 삭제 성공"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token. The frontend stores this token only in Zustand in-memory state. Do not store it in localStorage, sessionStorage, or cookies."
      }
    },
    "parameters": {
      "TodoId": {
        "name": "todoId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "할일 ID"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "잘못된 요청",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "인증 실패 또는 JWT 없음",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "리소스를 찾을 수 없음",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Conflict": {
        "description": "중복 이메일 또는 중복 카테고리",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "서버 오류",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "SignupRequest": {
        "type": "object",
        "required": [
          "email",
          "password",
          "displayName"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 255
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 100
          },
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "이름 또는 닉네임"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "token",
              "user"
            ],
            "properties": {
              "token": {
                "type": "string",
                "description": "JWT access token. Store only in Zustand in-memory state. Do not store it in localStorage, sessionStorage, or cookies."
              },
              "user": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        }
      },
      "AuthUserResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "UpdateMeRequest": {
        "type": "object",
        "required": [
          "displayName",
          "darkModeEnabled",
          "language"
        ],
        "properties": {
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "이름 또는 닉네임"
          },
          "darkModeEnabled": {
            "type": "boolean",
            "description": "사용자별 다크모드 사용 여부"
          },
          "language": {
            "type": "string",
            "enum": [
              "ko",
              "en",
              "ja"
            ],
            "description": "사용자별 화면 표시 언어"
          }
        }
      },
      "User": {
        "type": "object",
        "required": [
          "id",
          "email",
          "displayName",
          "darkModeEnabled",
          "language",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "displayName": {
            "type": "string"
          },
          "darkModeEnabled": {
            "type": "boolean",
            "description": "사용자별 다크모드 사용 여부"
          },
          "language": {
            "type": "string",
            "enum": [
              "ko",
              "en",
              "ja"
            ],
            "description": "사용자별 화면 표시 언어"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "HealthResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "example": "ok"
          }
        }
      },
      "CreateCategoryRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          }
        }
      },
      "CategoryResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Category"
          }
        }
      },
      "Category": {
        "type": "object",
        "required": [
          "id",
          "name",
          "isDefault",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "기본 카테고리는 null, 사용자 추가 카테고리는 생성 사용자 ID"
          },
          "name": {
            "type": "string",
            "example": "업무"
          },
          "isDefault": {
            "type": "boolean",
            "description": "기본 카테고리 여부"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateTodoRequest": {
        "type": "object",
        "required": [
          "title",
          "categoryId"
        ],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "description": "할일명"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "설명"
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "종료예정일"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "UpdateTodoRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "registered",
              "in_progress",
              "completed",
              "canceled"
            ],
            "description": "?? ??"
          }
        }
      },
      "TodoResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Todo"
          }
        }
      },
      "TodoListResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Todo"
            }
          }
        }
      },
      "Todo": {
        "type": "object",
        "required": [
          "id",
          "userId",
          "categoryId",
          "title",
          "status",
          "isCompleted",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "title": {
            "type": "string",
            "example": "주간 보고서 작성"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "금요일 회의 전까지 초안 작성"
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "2026-05-15"
          },
          "isCompleted": {
            "type": "boolean",
            "example": false,
            "description": "기존 호환 필드. status가 completed이면 true"
          },
          "status": {
            "type": "string",
            "enum": [
              "registered",
              "in_progress",
              "completed",
              "canceled"
            ],
            "example": "registered",
            "description": "?? ??"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string",
            "example": "요청 값이 올바르지 않습니다."
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    }
  }
}